refactor: 폴더 구조 변경 및 코드 정리
